[images] Write depfile for pave.sh and netboot.sh
Without this change, the build isn't able to figure out whether or
not pave.sh and netboot.sh are dirty, so it always considers these
steps dirty. With this change, pack-images.py will always write
out a depfile for these.

Test: Verified the build will recreate pave.sh and netboot.sh if
either of their dependencies, images.json or zedboot_images.json,
are changed.

Change-Id: I9d7c3521bc906e157f7dd948578070073e31c364