[wlan][mlme] Compare Join PHY and Association PHY
An expected behavior from the AP is if AP does not support
what the Client requested in AssociationRequest, the AP
should reject the Association. This CL tests if the
runtime association lives up to this expectation.

The failure of test is a simple warning in kernel log.
It is yet to be determined if the client needs to take
action and claim that the negotiation failed.

Test: Smoke with GoogleGuest
NET-1659 #done

Change-Id: I2bb47c257809a70494c498458974c78661efb07e