[fidl] Fix 0 sizing of non-primitive arrays
Arrays in FIDL can contain non-primitive element types, and in such
cases, the current version of the FIDL compiler ends up setting the size
of the array to 0. This is because the size calculation for arrays is
performed in the Consume pass, and IdentifierTypes' sizes are not
resolved at that point. This CL moves this calculation into the Compile
step, where the size of IdentifierTypes (and any other element type) can
be resolved.

FIDL-337 #done

Test: Compile FIDL file with array of both primitives and non-primitives
and ensure that the generated table contains the proper sizes.

Change-Id: I909d3f3ef2b3c33d25e9524a29267d5f4b378abe