[fidl] Method request/response as anonymous structs
Instead of representing a method's request and response as a special
case 'Message', using anonymous struct request and response.

To avoid any downstream changes, hiding anonymous structs from the JSON
IR, the C Generator, and the tables generator. However, we're adding an
'anonymous' tag in the JSON IR to allow backends to ignore these
structs, to eventually allow the compiler from emitting them, on the
path to represent methods in the JSON IR as they are represented in the
compiler.

Test: make tools -j12 HOST_USE_ASAN=true && ./build-x64/host_tests/fidl-compiler-test
Change-Id: I3e3df50d7d64fc07879bac55d573e525785e0a8b