[gn] Patch guess_layer.py for flower model.
`fx set` does not work after following the flower model checkout steps.

The import in a developer's manifest is always going to
point at a manifest in integration. Rather than rely on that
import's name (which can be omitted in the jiri import call)
rely on the name of the project that is <overidde>n in the
.jiri_manifest, since that is the new normal workflow.

DX-629 #comment

Change-Id: I405827c978fee0bae7c1f0634ae6ac46178a1aaf