[asan] Export __asan_default_options() so it has an effect
With -fvisibility=hidden, __asan_default_options is never seen by the
ASan runtime DSO and so has no effect.  Prevent using that switch for
this file to make sure it is seen as referenced at link time and gets
into the dynamic symbol table.

Also change the way we solve the issue for __scudo_default_options so
that it's compatible with host (i.e. non-Fuchsia) systems in case
anyone ever enables Scudo for host code.

Test: CQ
Change-Id: I86e548842d7f3a9ac232d03a668eee26b64906f0