[devmgr][dev] Use relative paths in load_firmware(), disallow absolute
Most users of load_firmware() already use only relative paths.  But
a few in the Zircon tree were using absolute paths.  The openat()
function ignores the given directory fd when given an absolute path
and just opens it as is, so devmgr was opening these absolute paths
as requested.

The paths in load_firmware() should never be absolute.  The driver
shouldn't know where things are stored, nor have any access to the
filesystem directly.  So change all the callers to use relative
paths, and make devmgr refuse to open absolute paths.

Test: CQ
Change-Id: I9580bb6da890fecf8cbb585a374fea5ae612e09d