[zircon] Fix channel data ordering
This change fixes a serious defect in the channel packet ordering
with respect to the current endpoint owner. That is, currently
the packet order does not reflect the owner order so that for
example the previous endpoint owner can inject a packet into the
channel after N packets of the current owner have been queued.

In effect this patch catches a  missing "bad handle" case for
channels. Other objects in effect suffer from this but it is only
critical for channels which have a single owner.

TEST =
  Unfortunately to fully test this you need to induce a
  thread race between 3 threads in two processes. Otherwise
  the current tests properly cover the non racy cases.

ZX-2695 #done

Change-Id: Ifa524d7bb24336ff337b93ac945f35cd84213776