[deb-sysroot-builder] Rename headers with conflicting case
Linux system headers contain files that only differ in case. This is
a problem for other systems that use case assensitive or insensitive
filesystems. To workaround this issue, we rename the conflicting files
by appending '_' to their name and then modify all headers that refer
to the old name to use the new one.

The same solution is used by Linaro for their Windows toolchain.

Change-Id: Ie5fee04a06d955b66517360c0f99c73d233ed977