[ledger] Move page usage db directory under <repo_path>/content.
Page usage information was stored under <repo_path>/page_usage_db/
This was problematic as deleting a repository directory couldn't be
done atomically: two separate directories (content/ and
page_usage_db/) needed to be moved to the staging area before being
deleted.

This CL solves this issue by moving <repo_path>/page_usage_db/ to
<repo_path>/content/page_usage_db/

It also fixes a bug, where the RepositoryInformation object was
deleted right before it was used in DeleteRepositoryDirectory:
OnVersionMismatch was calling LedgerRepositoryContainer::Detach()
leading to the deletion of UserSyncImpl which owns the
RepositoryInformation object.

Test: LedgerEndToEndTest.CloudEraseRecoveryOnInitialCheck,

Change-Id: Ic3089bf8cefcfab27d56f33e332e0a0def98c5f3