[ledger] Create new LevelDb instances in staging area.
If LevelDb initialization fails for some reason (e.g. is interrupted
or there is an I/O error), the created instance might be corrupted.
To avoid this, when creating LevelDb instances in LevelDbFactory we
now create them in a staging path and then move them to their final
destination.

Test: LevelDbFactoryTest.CreateGetDb

LE-617

Change-Id: Id762e259f13b3de2bc71e8e014ffaa67dc5e98bf