[fidl] Test fidlc for multi-pass correctness
This CL creates a new set of tests specifically for ensuring that the
operations during the multiple passes fidlc makes are in the correct
order. This should catch a class of bugs related to resolving
type-specific attributes during the consumption pass versus the compile
pass. Additional multi-file tests will be added to this.

Test: Run fidl-compiler-test.

Change-Id: I96c33d2597f5cea27aac2dce00f961eb61783af4