[debug registers][syscall] Add arm64 debug register support.
Enables userspace to read/write debug registers state into a particular
thread.
Added tests to verify validation and behaviour.

The next step is to track this information per-thread and add it to the
context switch. This is not a problem right now because we never enable
the global debug exceptions (bit KDE of mdscr_el1), which means that no
HW breakpoint exceptions will occur. This will be enabled once the
correct context switch and exception handling is working.

TEST=unit

Change-Id: Ib7215a04547590179de62a7c502389458e7f0dff