[fidl] Fix max out-of-line value for arrays
This fix propagates max out-of-line values from members into inline
arrays for FIDL. Previously, arrays would not account for the
out-of-line members of its elements and default to 0 as its out-of-line
count.

FIDL-343 #done

Test: Run existing FIDL tests. Run new multi-pass FIDL test stacked on
top of this change, with and without this change. With this change,
everything should pass. Without this change, the out-of-line checks will
fail.

Change-Id: I27cd03069f952324a6177348f0e669cf7b26de3b