[scenic] Remove escher::buffer offset and size
This change converts all users of escher::Buffer to use GpuMem as the mechanism for specifying a sub-region of memory to use for the buffer (i.e., by calling Allocate() on the GpuMem object). This removes the need to pass in separate offset and size parameters into Buffer.

In addition, it removes most accessors from Buffer, in preparation for a world where we allocate vkBuffer objects directly (through something like vk_mem_alloc.h) instead of through a GpuMem object.

The Sketchy client has been updated to use Buffer::New directly, instead of using the factory object, so that it a) can guarantee that the VMO backing the memory does not contain memory for other uses, and b) can extract the GpuMem from the factory function (now supported as an out parameter).

Finally, the BufferRange struct has been moved to impl/compute_shader.h, which is the lowest level interface that actually uses the struct directly.

tested = new test in buffer_unittest.cc, all existing tests still pass, manual running of sketchy client, and of video test apps

Change-Id: I1ee29a0190722b9e3fc751a1bddff870ff7fb71c