[aml][thermal] Disable AML thermal driver
The AML Thermal Driver is somehow throttling down the whole
system which is causing the system to perform poorly.

Not sure what the problem is yet but I'm disabling the thermal
driver until somebody has a chance to look at it.

Tested: Booted upper layers on Vim2 and observed choppy performance
then I applied these changes and booted again and observed that things
were smooth.

Change-Id: I3241bf0d5f415f76d34b5911ba139216c35326f0