[zircon][utest] Add test to check root hubs
This changeset adds a test to check whether the xhci roothubs are created
after the system boots up. It has many TODOs for now and is marked
skippable, as there are other dependent tasks that need to be done
before. First, we need a way to run only on hardware and second, we need to
be knowing which hardware it is running on. In a future changeset, I
will add a way to wait on the usb-device class creation. I am planning to
use the wait_for_device from ramdisk.cpp. I will do that in a separate
changeset as it needs to be moved out of ramdisk.cpp.

ZX-2929 # Added a skippable test to test bring up of root hub devices. Waiting on
IN-497 to make it non-skippable.

Test: fx full-build. runtests on Astro/vim2/pixelbook. Verified the tests failed
on Qemu without the "SKIP". With the SKIP tests passed.

Change-Id: Ieee5aa7482dfb254409c9779736de27aabe12780