[build][debug] No optimization for debug builds.
Previously we compiled debug builds with -Og. GCC defines this flag to
be "debugger-friendly optimizations". But reports from the Clang team
indicate that Clang remaps this to -O1. It also appears that or debug
code optimization is relatively unfriendly to debuggers.

This changes the debug mode to use -O0, removing optimization. This is
more in line with what people expect from a debug build.

Later we can add an intermediate level of optimization if desired, but
the longer-term plan is likely that people will compile their own
components using the SDK using the optimization level that suits their
task, and the system will always be a release build.

TEST=builds, run code in the debugger.

Change-Id: If06edac84e3e8d245c3d16383e723390b8e1bc39