[wlan][mlme][ap] Change Authenticating/AssociatingState constructors
Instead of transitioning into AuthenticatingState as soon as we receive
an authentication frame, make it a state where the MLME waits for an
authenticate response from SME instead. That is, only transition to
AuthenticatingState when the authentication frame is valid.

Also change AssociatingState to no longer take `aid` as a field.

Test: wlan_ap_unittest
Change-Id: I4e6cccba0271ba464cce1d8c1754d540a4d3ad2a