[devmgr] Switch devcoordinator to libasync
This is a relanding of 57372092ed90537eb815683b91f274fab1540e31 which was
reverted in e278a58e769d831f48b28e817023abadf28081c3 for causing
failures on the roller.  The only change between this patch and that are
naming changes.

This will get us things like future-scheduled events, which
would be useful for implementing backoff on driver binding attempts.

Test: Booted in qemu with ASAN enabled.  Booted on hardware, verified
system comes up and is responsive to hardware events.

Change-Id: Ia2fbc2c7fe1347184c1e3645daefdc1dec6f3ea0