[bt][snoop] Move and rewrite opts for snoop service
moved snooper from bluetooth/tools to bluetooth. This is part of
a migration to make the snooper a long lived service that stores
historical logs on an ongoing basis rather than an ephermeral tool.

Added unit tests for snooper and refactored to make it easier to test.
Updated sysmgr config and build files to include new snoop service
component.

NET-1257 #progress

TEST=unittests: fx run-test bt_snoop_bin_test_rustc

Change-Id: Ide113a62a7746469f3cfb764cd5688b21f25904e