[vim][thermal] Fix clocks not being set appropriately
Currently, the big/little cluster clocks are getting reset to their
lowest values of 100MHz (as observed from system_dash).  This is
causing the observed sluggishness.  The root cause is that in the
recent c++ rewrite of this driver, two things were missed:
  1. The core DVFS table indices were being confused with the device
     OPP indices.
  2. The available core OPPs were not being written into the ioctl-
     returned metadata.
This change fixes the above issues allowing the thermal daemon to set
the appropriate values for each of two power domains.  This resolves
ZX-2945.

Test: pave topaz onto device, observe snappier UI
Test: run system_dash and verify clock running at 1.296GHz@1.1V
Change-Id: Ic72a29b86e22d43ca321d9195d6e0270a0e09993