[debugger] Context aware get.
Now settings are showed according to each schema (System, Target,
Thread) and show the appropiate value according to the context (ie. if a
target overrides, it will show the overridden value). Feedback whether a
setting was overridden is pending.

Also moved the "level" setting from the store to the schema. A
SettingStore should not care about what "Setting level" it is, but
rather that is a schema concept. Changed the tests appropiatedly.

Setting will have to follow the same approach, as you could have an
override for a system setting at the target context.

TEST=manual

Change-Id: I7c90bdad9c2875e6871b6e3bc45f6ca67742f222