[media] Audio tools package group
To ensure that tools for debugging audio issues can be clearly
included in products when needed, this CL creates the package group
garnet/packages/tools/audio and ensures that all of these tools are
easily runnable from the command-line.

Building on the demonstration of available packages, this CL makes
the cmdline audio tools available in terminal.gni as well, so that
a simple 'signal_generator' demand-installs not only the binary,
but also the requisite service (audio_core) from prod/media_audio
(also not preinstalled).

Test: CQ, build, manual

Change-Id: I5439b41f97dce1a4b0cfe837d5e8902fa2ff725a