[ddk][pdev] Rewrite ddk::PDev as a wrapper around ddk::PDevProtocolProxy
The autogenerated protocol proxies are lacking some conveniance
functionality. In order to improve that we can extend them to perform
the functionality we want.

ddk::Pdev was a competing implementation to the ddk::PDevProtocolProxy,
but much of it's functionality was repetitive. By inheriting from the
former, we are able to slim down the implementation.

In addition, because the PDevProtocolProxy is just two pointers, there
was no need to make the type restricted to move only ref counting. We
allow it to be a copy type instead.

I2cChannel recieved similar treatement with regards to I2cProtocolProxy.

GpioPin provided no functionality on top of GpioProtocolProxy so it was
removed.

ZX-2923: #comment Rewrote ddk::PDev to provide conveniance functionality

Tested: Tested audio on vim2

Change-Id: Ieabbf730f0598eb77a39e0f42185f92d272be0c2