[mediaplayer] Distinguish packet vs buffer in FidlDecoder
This CL makes FidlDecoder, the wrapper for mediacodec decoders,
handle the case in which output buffer packet_index and buffer_index
values are distinct. This is required for VP9 support.

The previous approach to managing output buffers relied on packet and
buffer indices being identical. Initially, the outboard decoder was
given ownership of the buffers in the current set. When the decoder
produced an output packet, FidlDecoder would transfer ownership of
the buffer to the internal representation of the packet. When the
packet was deleted, the buffer would be recycled, and the recycle
was hooked to transfer ownership of the buffer back to the decoder.
This ownership tracking was unnecessary given that the outboard
decoder makes all output buffer allocation decisions.

In the new scheme, the outboard decoder still gets ownership of all
the output buffers, but that ownership remains in effect until that
set of buffers is no longer current. When the decoder produces an
output packet, FidlDecoder shares ownership of the buffer between
the outboard decoder and the new packet. When the packet is
deleted, the buffer reference is dropped and FidlDecoder tells
the outboard decoder that we are done with the output packet.

TEST: tested for regressions using mediaplayer_tests

Change-Id: I1cb145e39f5b03d0185350baeb863638df2d3849