[zxcrypt] Fix race condition between paver and zxcrypt binding
- During paving, the superblock of a zxcrypt partition is
written as a part of "zxcrypt::Volume::Create". This invokes
CommitBlock, which has a 'read -> check -> write' pattern.
- CommitBlock is simultaneously called by Unlock, while binding
the zxcrypt driver from the block watcher.
- When these race, the old and new copies of the superblock
are intermingled, and the old version of the superblock may
end up overwriting the newly written data (or vice versa,
where a new version of the superblock is written, but a driver
assuming the old version is still active).

To fix:

- Avoid binding zxcrypt automatically while paving.
- Avoid writing the superblock automatically during driver
initialization.

ZX-2963 #comment In Progress
Test: Manually paving on Astro, observing "/data" is persistent

Change-Id: I50f5ba58acf6d38ee7fa96cbcf64588859b6c920