[symbolize] Fix bug regarding leading zeros.
This patch removes leading zeros from numbers before parsing them.
This fixes an unknown bug where process ids and thread ids were
probably being handle correctly in the sense that they were unique
but were not being handled correctly in the sense that they were
being mapped to incorrect values. This also eleminates a possible
panic that was found in TC-273 by Todd where due to lossy uart an
'x' was dropped resulting in an '0x8' turning into '08' which
caused the panic because strconv.ParseUint interpeted that number
as octal due to the leading zero even though '8' is not an octal
digit (hence the panic).

Change-Id: Ifa65214622c5cbe7938f247064aebc5daa640996