[mediaplayer] remove deprecated test-only methods from Player
This CL removes Player.SetAudioRenderer and Player.SetReaderSource,
which existed only for the CQ tests. SetAudioRenderer is no longer
needed now that mediaplayer_tests is hermetic (fuchsia.media.Audio
is part of the synthetic environment in which mediaplayer runs).
SetReaderSource is no longer needed, because CreateReaderSource and
SetSource can do the same thing.

TEST: CQ test improvement - still passes
Change-Id: Ic2b49115ddc4b5967d46010a4bf6dd4449a8cbfa