[kernel][event] update comment on event_destroy()
This CL updates the comment for the event_destroy function. According to
the wait_queue_destroy() function[0], if there are threads waiting for
the queue that is being destroyed, it will panic.

It looks like it was just a stale comment. The first implementations of
the wait_queue used to wake up all threads. The behavior was changed in
Nov 2016[1] (this is back when wait_queue lived inside thread.c).

[0]:
https://fuchsia.googlesource.com/zircon/+/f0441ea7af4261943ba0d5a07cfa19a5489f4746/kernel/kernel/wait.cpp#450
[1]:
https://fuchsia.googlesource.com/zircon/+/9b052e953a9078e093384072c557202c983c1b6e

TEST=comment change. No test.

Change-Id: Ib99025e0002ade6006eb628d948f1cc491e902e5