[guestmgr][vmm] Enable persistent disk.
This adds a 'block_devices' field to the guest LaunchInfo structure
that applications can use to provide additional block devices to
the vmm. The application must initialize this file to some thing the
target guest can work with. For guests that can handle unformated
devices, an empty file will suffice.

This also removes support for the 'guid' and 'type-guid' block options.
These existed to work-around the lack of a persistent disk on minfs. If
a guest does want to pass through a physical partition, this is still
possible by passing in fuchsia.io.File for that partition, but the feature
will not be supported by the vmm directly.

This allows us to remove access to '/dev/class/block' as well as the
'shell' feature from the vmm's sandbox config, both of which were only
used to allow the vmm to open() block devices in different scenarios.

Test: Launched biscotti, verified file system modifications persisted
      across reboots.
      Launched linux_guest.
Change-Id: I6a853c9984b57aecbd3640e6b364d4889971d76f