[tee][optee] Add TEEC Operation Value handling
TEE Client Operations allow for two different kinds of command
parameters: Value and Buffer types. This CL adds supports for the
former. TEEC_Value parameters are just a pair of 32 bit unsigned
integers that can be used as either an input, output, or both. They are
intended to be an optimization for small fields to avoid buffer
allocation.

Test: tee-test on Astro
Change-Id: I8f2c5cb8b4d5629f10c1f90befdd21cae354d794