[tee][optee] Add TEEC Operation Buffer handling
TEE Client Operations allow for two different kinds of command
parameters: Value and Buffer types. This CL adds supports for the
latter. Unlike Value parameters, the Buffer parameters are fairly
complex to support a variety of different use cases.

The TEE Client API provides two interfaces for reserving Shared Memory:
TEEC_RegisterSharedMemory and TEEC_AllocateSharedMemory. The former is
to register an existing user buffer with the TEE, whereas the latter
requests that a buffer be provided by the API. Due to current system
restraints and a lack of use cases, the Register variant will remain
unimplemented for now.

Once a TEEC_SharedMemory object has been allocated, it can then be used
in TEEC_Operations. TEEC_Operation allows for a variety of different
buffer types, called memory references. Temporary memory references do
not have a TEEC_SharedMemory parent object and are only valid for the
duration of a command. Registered memory references refer to an
allocated TEEC_SharedMemory object as either a whole or partial
reference. A whole memory reference implies that the whole
TEEC_SharedMemory object can be used, whereas partial memory references
restrict the portion of the TEEC_SharedMemory object according to the
provided offset and size.

Test: tee-test on Astro.
Change-Id: Ifa4c4565a5198dd6eceed3573510bffe534231d3