[audio_core] Fix fzl::RefCountedVmoMapper dependency.
fzl::RefCountedVmoMapper has been removed from libfzl.  The idea is
that this is a trivial combination to create, and has only one user in
the system today (audio_core).  Because of this, the suggestion is
that users who want to manage VmoMappers with ref counted semantics
can just roll their own version.

Do exactly this for the audio_core.  Add the definition to utils.h and
update the curent users of the code.

Tests: Build and signal_generator spot check on Intel NUC
Change-Id: I8f8996dc9d07e61aa9da02a54282820884e1dbe3