[packages] Add prebuilt_package() template to integrate prebuilt bits
This adds a prebuild_package() GN template which is like a package()
except instead of referring to products built during the build it
points to an archive generated by 'pm archive'. The template expands to
rules that open the archive up using 'pm expand' and then generate the
response files we use to assemble system images and OTA directories.

Change-Id: I2e5a5a574df0b03574ff262d9ee2809ae4b801c9