[bt][host] Remove Profile Disconnect
Detecting a profile client disconnect is not well defined, so it is left
to the profile implementation to decide, based on the L2CAP sockets
that have been disconnected.

Sockets will disconnect when the channels close from the remote end,
which should provide feedback necessary for detecting disconnections.

Remove Profile.DisconnectClient() similarly - profiles can implement
this by closing the sockets that have been connected to them, which
will close the associated L2CAP channels.

Test: fx build - this is a strict removal
NET-1824 #done

Change-Id: I762774853164f9fb0c8b3a8bddd6b47d2205498b