[audio_core] Add ForEachSourceLink
This CL adds the ForEachSourceLink iterator to bookend the recently-
added ForEachDestLink, to be used by the AudioCapturer.

This CL also moves all three of the ForEach iterators to now pass
const shared_ptr<AudioLink>& , rather than AudioLink*.

Manual testing includes signal_generator and wav_recorder (with and
without --loopback) on a device with multiple audio inputs and
outputs, observing that (for example) SetGain calls are distributed
to all the appropriate links.

Test: build, CQ, manual

Change-Id: I98ac28406534db2992b27822379525e7936da327