[inspect] Allow operations on invalid object dir
It is occasionally useful for tests to not have a full startup context
provided to the code under test. In these cases we want to be able
to provide a compatible ObjectDir implementation that doesn't require
manually checking that it is valid. This change allows ObjectDir to be
used even if it does not point to a valid object, suppressing the effect
of the operations.

TEST=inspect_integration_test

CF-204: #progress

Change-Id: I0a7d8c291d4b85d84ea3be1ac5f15a099bcaa2bc