[input] Allow non-modifier key presses through CaptureKeyboardEventHACK
Currently, the method to capture keyboard events exposed to higher layers
relies on kbd.modifiers & event.modifiers evaluating to a truthy value.
However, if you need to listen to events without a modifier (in my case, I wanted
arrow key events) 0 & 0 is false.
This CL changes the expression to allows events only if their modifier values are exactly equal.

Additionally, to facilitate listening to key presses without a codePoint (in particular, arrow keys)
this CL checks that event.hid_usage == kbd.hid_usage for events with a codePoint of 0

This second concern could be broken into a separate CL. I think that the behavior of treating the codePoint of 0
specially is not great - potentially it would be nice to be able to pass a null value for codePoint or hidUsage instead.


test: manual verification in user shell
Change-Id: I0b8472c737d395fe48843a8c56317070529272c0