[cobalt-client]: Timer update.
Previously I had the assumption that we would know what we are logging
a priori, which is not always true. For example, when Operation A
has multiple branches and different branches map to different event
codes, we dont know which branch we are gonna pick until it is picked.
At this point binding the timer to a specific histogram, is less
useful than simply providing a fs::Ticker equivalent.

For this reason, removed the histogram as part of the timer,
and we can do something like:
Timer timer(is_collecting);

.....
    hitogram.Add(timer.End());
TESTS= cobalt-client-test

Change-Id: I93731f983112b0a061130c8f6f7bab9d7f98cffc