[ephemeral] Remove ability to launch binaries as components
This change takes away the ability to provide a URL to `run` or
`Launcher.CreateComponent` that resolves to an arbitrary binary.

This URL can take a couple forms:
- A full path to a binary under /pkgfs, /system, or /boot
- A "naked" binary name that is resolved to /system/bin/...

This feature is a security hole and is not compatible with having cmx
for every component. More immediately, it causes bugs with
auto_update_packages=true. Specifically, the package updating loader
does not attempt to find a matching binary if the package update fails
because the package does not exist. We could fix this behavior, but
since the feature is undesirable in the first place, it is better to
remove it.

To accomplish this, this change introduces many cmx files for
components and tests that were formerly cmxless.

service_startup_test and package_updating_loader are also modified to pass
when service_startup_test is run with auto_update_packages=true.

CF-129 #comment Removed ability to launch arbitrary binaries as components
CF-156 #comment Removed ability to launch arbitrary binaries as components
CF-169 #comment Removed ability to launch arbitrary binaries as components

CF-169 #comment Removed ability to launch arbitrary binaries as
TESTED=Manually found and fixed uses of `run` or `CreateComponent` on

non-fuchsia-pkg URLs, and updated tests as necessary

Change-Id: I989cb064bd037a3e66c0b7d95d6cd6d8a858a431