[mdns] rename mdns_service->mdns, mdns->mdns_util
This CL is identical to
https://fuchsia-review.googlesource.com/c/garnet/+/220780, except
that the mdns_service package file has been retained so this is a
soft transition (the garnet roller won't break). The mdns_service
package file will be removed later.

From the original:

This CL renames the mdns_service binary to 'mdns'. The old mdns
binary, which is a utility, has been renamed mdns_util.

TEST: no behavior change
Change-Id: I7d4ba98572beb82255cf00fb801a485f1ee0cef0