[cobalt-client]: Remove forward declarations.
Due to fbl::unique_ptr lack of support for incomplete types,
there is not point in attempting to use PIMPL, to reduce the
user needed dependencies. This will mean that uses will have to
import the fidl dependencies additonally from the library itself.

Eventually, with GN, this should never be spoken of again.

TEST=cobalt-client-test

Change-Id: I9e37f0bdcc4c0f37a9b890472892690570581645