[debugger] Move backtrace to ThreadRecord.
A ThreadRecord IPC object contains the thread's current state.
Previously this did not include backtrace information and the places
that needed a backtrace (mainly the exception notification) sent it
separately.

This produces missing backtraces if the thread ever becomes suspended
without an exception.

This patch adds the backtrace to the thread record so the minimal (2
element) backtrace is sent whenever we report the state of a suspende
thread.

TEST=unit,manual

Change-Id: I76eb3a8ee7e4f4271a9a46bf8cc8a911d3e4aec7