[kernel][pmm][unittest] move the pmm unittests into their own group, reenable one
The unittest that grabs a few pages at once should be okay, it's only
64KB. The other one that explicitly overallocates was the dangerous one
that was disabled a while back.

Remove 'tests' from the unittest name, to be consistent with others.

Tested: build on x86/arm, run 'k ut pmm' also 'runtests' to observe
this test gets run then as well.

Change-Id: If3cfb291b7cab025cbfb4a6e8f77e551bd7bbe6a