[debugger] Send thread state with backtrace.
Replaces the backtrace request for a more general "thread state"
request. The thread state includes the backtrace, and also includes to
exception state of the thread and other statistics. This is important
because these things are related. Previuosly it was possible to get an
old thread state (like "running") and a new backtrace, while normally
those two things wouldn't go together.

With this change, the thread's metadata and frames are consistent (as of
the request time).

Most of the change is just renaming the request functions. A new
function FillThreadRecord is added to the DebuggedThread to populate the
full thread state.

Some ifdefs around a thread state Kernel API update are removed.

The documentation and some user messaging is clarified around when
backtraces are available. Previously whether this was "blocked" or
"suspended" was ambiguous or confusing.

The stack amount (full vs minimal) are made explicit on the ThreadStatus
structure rahter than being implicit based on the message type, and
passed around on both ends.

TEST=unit,manual

Change-Id: I575b8b1f52f0d383b7f747f36ec0b10a36a2ca7f