[profile] Add simple profile object service
This new service hosted in svchost vends profile objects which are used
to set thread priority.

This is not the final form of the service, it is just a necessary step
to be able to remove zx_thread_set_priority().

This CL is based on cpu's at
https://fuchsia-review.googlesource.com/c/zircon/+/229193.

ZX-940 #comment [profile] Add simple profile object service

Test: new test `/boot/test/sys/profile-svc-test`

Change-Id: Ic40fbe16c5dd9d0e65ddd5531a195801cbb2fa23