[mediaplayer] Count missing bytes.
FindOrCreateHoles is necessarily called before CleanUpExcept to determine a
budget, but CleanUpExcept will likely invalidate the holes returned by moving
the underlying iterators around, which is easy to mess up.

This function just counts the missing bytes so users aren't required to
think about the hole references becoming invalid.

If only C++ had some way of checking the lifetimes of borrows of a
resource...

TEST: fx run-test mediaplayer_demux_tests
Change-Id: I8549fc85360c80f1166fb9b255154d0d8be4c79e