[debugger] Quit the debug agent from zxdb.
Sends a new IPC message to quit the debug agent.

Fixed a bug where the message loop would be deleted before the released
resources would be using them. This is done by extracting the message
loop from the connection management.

Fixed a bug where deleting the connection when trying to disconnect
would crash the client.

Added a flag to disconnect the debug agent on client exit.

TEST=manual

Change-Id: I47be2e6119d8cf0b63778bf40230f6d71d38f531