[sshd-host] sshd-host replaces listen
This is a small step in the direction of componentizing sshd, in that it
actually puts all of the sshd related items into a single component.

It still uses chrealm, as users want to be able to access all services from
the root realm in programs that are spawned from subshells.

sshd itself can't yet be a standalone component because it currently accepts
the connection socket over stdin, and components do not have stdin.

sshd-host takes on the responsibility of running `hostkeygen` once at
startup, in order to ensure that host keys are available. `hostkeygen` could
be further sandboxed in the future, but for now the program is extremely
simple, so that chore best remains until we have a better way to handle these
kinds of binaries in future.

Test: build & ensure that ssh to target still works.
Change-Id: Id59ea9652b379ecdb8b0b35daaa62f92af52e9f3