[virtio][wayland] Move virtio_wl out of process.
This change splits the virtio_wl device and moves it out of process.
Small changes were made to the device to adapt to the new model.
Specifically, the async waits on the queues have been replaced with
logic triggered in the NotifyQueue method. When new command buffers
are received, they're processed immediately. When a new output buffer
is received, we flush any 'ready' VFDs immediately.

MAC-173 #done

Test: weston-simple-shm still works.
Change-Id: I7b4af109469cf9c3917f53a5b6bde476dbd3bbe1