[images] publish blobs for "available" packages
All blobs for all built packages need to be published to the amber
repository. Prior to this change only blobs that were members of the "system
image" were being published.

This change also re-arranges the build rules in order to make use of the
package tool output `blobs.json` in order to avoid repeated computation of
merkleroots, significantly speeding up final build steps.

Bug: BLD-316 #done
Test: add kitchen_sink to available_packages, then `fx build-push` and observe successful completion of all of kitchen_sink being installed.
Change-Id: I49ed341af83a45b3237bb4e8a733fe779b63b15f