[cobalt-client]: Remove option for pre allocating.
No longer need to define max_histograms/max_counters.

Use a wrapper class to dereference a pointer to the vector
and an index in the vector, ElementView.

Histogram and Counter used to have a pointer to their respective
internal implementations(Remote and local in the future). This meant
that the vector that contained them was not allowed to reallocated,
which is why we requested the amount of each before allocation.

Instead now we prevent Collector from being moved, and provide only
heap allocated Collector. This allows to have a pointer to the vector
and index within the vector, fetching on demand.

Due to lack of support for immutable vector mutable elements, or
vecotr view (std::span) we provide an equivalent of for a single
element, eq of span of size 1, ElementView.

TEST=cobalt-client-test

Change-Id: Ie18af1b3c6927cfe41282d8ac951fce2c1dff181