[devmgr] Fix race between async loop and device remove
device_remove() may be called from a different thread than the async
loop.  Because of that, we need to avoid mutating the async connection
data structure from device_remove().  This patch reworks the shutdown
logic to avoid unsafe accesses.

ZX-2961 #comment Fixes the non-proxy datarace

Test: Ran /pkgfs/packages/wlan-hw-sim-tests/0/test/wlan-hw-sim_bin_test_rustc
repeatedly.  Previously would trigger a devhost crash roughly 1 in 5
times.  Did not see anymore devhost crashes with this patch, except for
the discovery of a bug in the wlan driver (NET-1860).

Change-Id: Ib48d058b0f5f4d5b1958d6ea688c67ae1e53b6d1